chore(orchestration): consolidate nine helpers into megaplan/orchestration/
Move evaluation, phase_result, iteration_pressure, verifiability,
tiebreaker, parallel_critique, progress, feedback, and audit (renamed to
plan_audit to disambiguate from megaplan/audits/) into a new
megaplan/orchestration/ package. Atomically rewrite all internal callers
across megaplan/ and tests/ (imports, string-form mock targets, the
phase_result logger name, and tiebreaker's deferred audit imports) per
U2 (no compatibility shims). Package __init__ intentionally performs no
eager imports to avoid partial-import / circular-import drift with
megaplan._core and megaplan.store.

Top-level re-exports in megaplan/__init__.py (build_orchestrator_guidance,
build_gate_signals, compute_plan_delta_percent, compute_recurring_critiques,
flag_weight) continue to resolve at the same megaplan.* paths.
