+++ tests/test_phase_result.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from megaplan.phase_result import (
+from megaplan.orchestration.phase_result import (
     BlockedTask,
     Deviation,
     ExitKind,
